Edits to Zoom SDK v5.0.24433.0616 to fix issues and enable to work on Windows 10
Changes:
+ Created Gitignore based on https://github.com/github/gitignore/blob/master/VisualStudio.gitignore to remove build output

+ Support for VS 2019:  Upgraded project files to use the latest version of the C++ VS runtime and fixed it so we can build with Visual Studio 2019 Community Edition.  It looks like they used VS 2015 before.  Also changed the Windows SDK targeting from 8.0 to 10 (latest, using 10.0.18362.0 on my build machine) and upgraded the platform toolset to v142.

+ Processor architecture mismatch, targeted MSIL, but requires x86.  Updated configuration manager to build all projects (was not set to build the wrapper) and set the active solution platform to x86 + release.  Release just to make sure performance is good (it worked with both)

+ Before needed to add CSHARP_WRAP as an additional compiler flag, but they fixed that in the last SDK update

+ Disable precompiled headers (NotUsing).  Note that the VS C/C++ project UI doesn't show this properly, so examine the vcxproj file.  This slows down the build immensely but was not able to get stdafx.h generating properly.  This output will be checked in/not built often to mitigate this.

Notes:
Output is done through custom actions to zoom-c-sharp-wrapper\bin.  The entire folder must be copied.  It includes the sample app, but that's ok, as we won't run it.

bin\zoom_sdk_dotnet_wrap.dll was committed in the original repo, it should not have been.  The build modifies it.  Just ignore this modification, we should try to minimize changes from the original version.

SDK upgrade:
Previously I was using v4.6.21666.0428.  There was a major upgrade because of Zoom's "security" push during 2020, and they changed how SDK apps authenticate.  However, the new JWT route does not work, but using the JWT meeting join route and the v4 way of an SDK user worked.  This requires some changes in MeetingRecorder to change the join user type from SDK_UT_APIUSER to SDK_UT_WITHOUT_LOGIN and the join parameter class from JoinParam4APIUser to JoinParam4WithoutLogin.  Future upgrades should keep an eye on this to make sure it doesn't break us.  New SDK version is v5.0.24433.0616.  Also ensure the reference in ZoomController to the wrap DLL is updated properly.

Testing:
+ Took the output in bin/* and copied to working directory for meeting recorder.  Ensured it could build and join a Zoom meeting successfully.

Changes to app:
+ The release notes mention the SDK now supports JWT-based authentication.  It neglects to mention that they removed support for joining as an SDK user.  In zoom_sdk_c_sharp_wrap\meeting_service_dotnet_wrap.cpp, CMeetingServiceDotNetWrap::Join only supports joining as a normal user or with SDK_UT_WITHOUT_LOGIN.  MeetingRecorder updates were needed.  See https://github.com/zoom/zoom-sdk-windows/blob/master/CHANGELOG.md for the details on JWT join (again, not working) -- for future reference.